Revert "Type fix"
This reverts commit ab3b9304d16234c599d40f34ddcdb5fdcdb702f0.

+++ src/General.cpp
@@ -2,8 +2,8 @@
 
 vector<string> tokenize(string s, string del = " ")
 {
-    int end = (int)s.find(del);
-    int start = 0;
+    size_t end = s.find(del);
+    size_t start = 0;
 
     vector<string> columns;
